+++ passes/opt/clean/cells_all.cc
@@ -19,7 +19,7 @@
 
 #include "kernel/ffinit.h"
 #include "kernel/yosys_common.h"
-#include "passes/opt/opt_clean/opt_clean.h"
+#include "passes/opt/clean/opt_clean.h"
 
 USING_YOSYS_NAMESPACE
 PRIVATE_NAMESPACE_BEGIN
